docs: comprehensive documentation restructure and optimization
- Add API reference (docs/api.md) and performance guide (docs/performance.md)
- Restructure CHANGELOG to Keep a Changelog format with version tracking
- Optimize .kiro specs with tables, status indicators, and traceability matrix
- Create custom Jekyll layout (_layouts/default.html) with responsive design
- Enhance GitHub workflows with separate lint/test/docs jobs
- Update index.md as documentation homepage with quick start guide
- Improve CONTRIBUTING.md with quick reference tables
- Enhance CLAUDE.md with architecture overview and common tasks
- Fix code formatting (ruff format) across all Python files
- Add divide-by-zero protection in CUDA kernels
- Add empty tensor validation in Python bindings

中文说明：
- 新增 API 参考文档和性能调优指南
- 重构 CHANGELOG 为标准格式，添加版本追踪
- 优化 .kiro 规格文档，添加表格和状态指示器
- 创建自定义 Jekyll 布局，支持响应式设计
- 增强 GitHub workflows，分离 lint/test/docs 任务
- 更新文档首页，添加快速开始指南
- 改进贡献指南，添加快速参考表
- 修复 Python 代码格式问题
- 添加 CUDA 内核除零保护
- 添加空张量验证

+++ src/tiled_attention.cu
@@ -56,8 +56,8 @@ __global__ void tiled_attention_kernel(
         row_sum[m] = 0.0f;
     }
     
-    // Pointer offsets
-    int offset = (batch_idx * num_heads + head_idx) * seq_len * head_dim;
+    // Pointer offsets (use int64 to avoid overflow for large tensors)
+    int64_t offset = (static_cast<int64_t>(batch_idx) * num_heads + head_idx) * seq_len * head_dim;
     const T* q_ptr = Q + offset;
     const T* k_ptr = K + offset;
     const T* v_ptr = V + offset;
